+++ SleepHunter/Converters/NumericConverter.cs
@@ -25,12 +25,12 @@ namespace SleepHunter.Converters
             {
                 if (integerValue >= MillionsThreshold)
                 {
-                    var fractionalMillions = integerValue / (double)MillionsThreshold;
+                    var fractionalMillions = integerValue / 1_000_000.0;
                     return $"{fractionalMillions:0.0}m";
                 }
                 else if (integerValue >= ThousandsThreshold)
                 {
-                    var fractionalThousands = integerValue / (double)ThousandsThreshold;
+                    var fractionalThousands = integerValue / 1_000.0;
                     return $"{fractionalThousands:0}k";
                 }
             }